perf

Struct CommonOpt

pub struct CommonOpt {

    pub send_buffer_size: u64,
    pub recv_buffer_size: u64,
    pub conn_stats: bool,
    pub keylog: bool,
    pub initial_mtu: u16,
    pub no_protection: bool,
    pub initial_rtt: Option<u64>,
    pub ack_frequency: bool,
    pub cong_alg: Option<CongestionAlgorithm>,
    pub stream_receive_window: Option<u64>,
    pub receive_window: Option<u64>,
    pub send_window: Option<u64>,
    pub max_udp_payload_size: u16,
    pub qlog_dir: Option<PathBuf>,

}

Fields
send_buffer_size: u64

Send buffer size in bytes

This can use SI suffixes for sizes. For example, 1M will request 1MiB, 10G will request 10GiB.

recv_buffer_size: u64

Receive buffer size in bytes

This can use SI suffixes for sizes. For example, 1M will request 1MiB, 10G will request 10GiB.

conn_stats: bool

Whether to print connection statistics

keylog: bool

Perform NSS-compatible TLS key logging to the file specified in SSLKEYLOGFILE.

initial_mtu: u16

UDP payload size that the network must be capable of carrying

no_protection: bool

Disable packet encryption/decryption (for debugging purpose)

initial_rtt: Option<u64>

The initial round-trip-time (in msecs)

ack_frequency: bool

Ack Frequency mode

cong_alg: Option<CongestionAlgorithm>

Congestion algorithm to use

stream_receive_window: Option<u64>

Maximum number of bytes the peer may transmit without acknowledgement on any one stream before becoming blocked.

This can use SI suffixes for sizes. For example, 1M will limit to 1MiB, 10G will limit to 10GiB.

receive_window: Option<u64>

Maximum number of bytes the peer may transmit across all streams of a connection before becoming blocked.

This can use SI suffixes for sizes. For example, 1M will limit to 1MiB, 10G will limit to 10GiB.

send_window: Option<u64>

Maximum number of bytes to transmit to a peer without acknowledgment

This can use SI suffixes for sizes. For example, 1M will limit to 1MiB, 10G will limit to 10GiB.

max_udp_payload_size: u16

Max UDP payload size in bytes

qlog_dir: Option<PathBuf>

qlog output directory

Alternatively you can set the QLOGDIR environment variable.

Implementations

impl CommonOpt

pub fn build_transport_config(&self, name: &str) -> Result<TransportConfig>

pub fn bind_socket(&self, addr: SocketAddr) -> Result<UdpSocket>